requirements.txt
python-docx==0.8.6
pyPdf2
pypandoc
